DTO 기능 확인 및 수정
