+++ src/test/java/racingcar/RacingDTOTest.java
@@ -2,9 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
-import java.util.List;
+import racingcar.domain.racing.RacingDTO;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -17,19 +15,8 @@ public class RacingDTOTest {
         int attempts = 3;
 
         String test ="aim, test, tenu";
-        RacingDTO racingDTO = RacingDTO.of(test, attempts);
-
-        String[] testArr = {"aim", "test","nony"};
-        RacingDTO racingDTO1 = new RacingDTO(testArr,attempts);
-
-        List<String> testList = new ArrayList<>();
-        testList.add("aim");
-        testList.add("tatoo");
-        RacingDTO racingDTO2 = RacingDTO.of(testList, attempts);
-
+        RacingDTO racingDTO = new RacingDTO(test, attempts);
         assertThat(racingDTO.getAttempts()).isEqualTo(attempts);
-        assertThat(racingDTO1.getAttempts()).isEqualTo(attempts);
-        assertThat(racingDTO2.getAttempts()).isEqualTo(attempts);
 
     }
 }
